# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Pomodoro App",
        "title": "Pomodoro API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/sessions/active": {
            "get": {
                "description": "Gets all active sessions for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pomodoro Session"
                ],
                "summary": "Get Active Sessions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/end/{id}": {
            "post": {
                "description": "Ends an active pomodoro session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pomodoro Session"
                ],
                "summary": "End Pomodoro Session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Skip the session",
                        "name": "is_skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/history": {
            "get": {
                "description": "Gets session history for a user with pagination and filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pomodoro Session"
                ],
                "summary": "Get Session History",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session Type (focus, short_break, long_break)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Session Status (active, completed, skipped)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start Date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End Date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of sessions per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/start": {
            "post": {
                "description": "Starts a new pomodoro session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pomodoro Session"
                ],
                "summary": "Start Pomodoro Session",
                "parameters": [
                    {
                        "description": "Pomodoro Session Data",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSessionDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.CreateSessionDTO": {
            "type": "object",
            "required": [
                "duration",
                "type",
                "user_id"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "ended_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.SessionStatus"
                },
                "task_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.SessionType"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.SessionStatus": {
            "type": "string",
            "enum": [
                "active",
                "break",
                "skipped",
                "completed"
            ],
            "x-enum-varnames": [
                "SessionActive",
                "SessionBreak",
                "SessionSkipped",
                "SessionCompleted"
            ]
        },
        "model.SessionType": {
            "type": "string",
            "enum": [
                "focus",
                "short_break",
                "long_break"
            ],
            "x-enum-varnames": [
                "Focus",
                "ShortBreak",
                "LongBreak"
            ]
        },
...
    }
}